requirements.txt
typer>=0.12
python-dotenv>=1.0
rich>=13.7
python-telegram-bot>=21.4
jinja2>=3.1
peewee>=3.17
psutil>=6.0
pyyaml>=6.0
jsonpath-ng>=1.6
aiofiles>=23.2